+++ feature/cat-editor/src/main/java/com/dede/android_eggs/cat_editor/CatSvgCodeDialog.kt
@@ -142,13 +142,15 @@ private fun buildXmlAnnotatedString(
     colorRenderChar: Char = '■', // fill='#FF0000' will be rendered as fill='■#FF0000'
     transparentColorRenderChar: Char = '□'
 ): AnnotatedString {
-    // "(?<name>[\\w_-]+)=[\"'](?<color>#\\w{3,8})[\"']")
-    val colorAttributeRegex = Regex("([\\w_-]+)=[\"'](#\\w{3,8})[\"']")
+    // https://stackoverflow.com/questions/27834463/android-java-regex-named-groups/27834803#27834803
+    val colorAttributeRegex =
+        //Regex("(?<name>[\\w_-]+)=[\"'](?<color>#\\w{3,8})[\"']")
+        Regex("([\\w_-]+)=[\"'](#\\w{3,8})[\"']")
     var offset = 0
     val xmlBuilder = StringBuilder(xml)
     colorAttributeRegex.findAll(xml).forEach {
-        val gColor = it.groups[2]
-        val gName = it.groups[1]
+        val gColor = it.groups[2]// color
+        val gName = it.groups[1]// name
         if (gColor == null || gName == null) {
             return@forEach
         }
@@ -168,8 +170,9 @@ private fun buildXmlAnnotatedString(
     val tagsRegexValue = xmlTags.joinToString(prefix = "(", separator = "|", postfix = ")")
     val tagStartRegex = Regex("<$tagsRegexValue?")
     val tagEndRegex = Regex("/?$tagsRegexValue?>")
-    // "(?<name>[\\w_-]+)(?<attrSign>=[\"'](?<color>$colorRenderChar?)(?<value>[^'\"]*)['\"])"
-    val attributeRegex = Regex("([\\w_-]+)(=[\"']($colorRenderChar?)([^'\"]*)['\"])")
+    val attributeRegex =
+        //Regex("(?<name>[\\w_-]+)(?<attrSign>=[\"'](?<color>$colorRenderChar?)(?<value>[^'\"]*)['\"])")
+        Regex("([\\w_-]+)(=[\"']($colorRenderChar?)([^'\"]*)['\"])")
 
     fun AnnotatedString.Builder.addStyle(style: SpanStyle, range: IntRange) {
         addStyle(style, start = range.first, end = range.last + 1)
@@ -190,19 +193,19 @@ private fun buildXmlAnnotatedString(
         val attrSignStyle = SpanStyle(color = attributeColor)
         val attrValueStyle = SpanStyle(fontWeight = FontWeight.Medium)
         attributeRegex.findAll(modifiedXml).forEach {
-            val gName = it.groups[1]
+            val gName = it.groups[1]// name
             if (gName != null) {
                 addStyle(attrNameStyle, gName.range)
             }
-            val gAttrSign = it.groups[2]
+            val gAttrSign = it.groups[2]// attrSign
             if (gAttrSign != null) {
                 addStyle(attrSignStyle, gAttrSign.range)
             }
-            val gValue = it.groups[3]
+            val gValue = it.groups[3]// value
             if (gValue != null) {
                 addStyle(attrValueStyle, gValue.range)
 
-                val gColor = it.groups[4]
+                val gColor = it.groups[4]// color
                 if (gColor != null) {
                     val color = gValue.value.toColorOrNull()
                     if (color != null && color.alpha > 0f) {